Rename miscelleneous.py to misc.py
Provides also the old naming for backward compatibility
